core/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.3
email-validator==2.3.0
python-jose==3.4.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pytest==8.3.5
